+++ src/hapi_schema/utils/constraints.py
@@ -43,6 +43,14 @@ def non_negative_constraint(
     return CheckConstraint(sqltext=sqltext, name=f"{var_name}")
 
 
+def rate_constraint(
+    var_name: str = "rate",
+) -> CheckConstraint:
+    """Require a column to be between 0.0 and 1.0."""
+    sqltext = f"{var_name} >= 0 AND {var_name} <= 1.0"
+    return CheckConstraint(sqltext=sqltext, name=f"{var_name}")
+
+
 def reference_period_constraint() -> CheckConstraint:
     """reference_period_end should be greater than reference_period_start"""
     sqltext = "reference_period_end >= reference_period_start "