+++ changelog.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## (next release)
+
+Miscellaneous fixes for pipelines
+
+### Changed
+- added from\_cods column to location, admin1, and admin2
+
+
 ## 0.8.1
 
 Implemented the Food Prices subcategory.