+++ changelog.md
@@ -5,16 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.8.3
+## 0.8.4
 
 ### Added
 - New constraints: rate\_constraint
+- refactor DBPovertyRate to track rates rather than population
+
+## 0.8.3
 
 ### Changed
 - fixed WFP commodity enum
 - change FLOAT to DECIMAL in funding, national\_risk, and food\_security
 - remove ALL ("*") from PovertyClassification enum
-- refactor DBPovertyRate to track rates rather than population, and link to location (with admin1\_name as a text field)
 
 ## 0.8.2